Restructured solving modules
