--- src/database/object/mod.rs
@@ -1,12 +1,0 @@
-//! # Database Object Module
-//!
-//! This module contains the definitions for different database primitives such
-//! as records and tables that are used in some database implementations.
-//!
-//! #### Authorship
-//!
-//! - Max Fierro, 2/24/2024 (maxfierro@berkeley.edu)
-
-pub mod record;
-pub mod schema;
-pub mod table;
