--- src/interface/terminal/mod.rs
@@ -1,12 +1,0 @@
-//! # Command Line and Terminal Module
-//!
-//! This module defines a standard means of interaction with GamesmanNova
-//! by assuming UNIX utilities for interfacing with a terminal.
-//!
-//! #### Authorship
-//!
-//! - Max Fierro, 4/6/2023 (maxfierro@berkeley.edu)
-
-/* MODULES */
-
-pub mod cli;
