+++ src/solver/strong/acyclic.rs
@@ -1,0 +1,160 @@
+//! # Strong Acyclic Solving Module
+//!
+//! This module implements strong acyclic solvers for all applicable types of
+//! games through blanket implementations of the `acyclic::Solver` trait,
+//! optimizing for specific game characteristics wherever possible.
+//!
+//! #### Authorship
+//!
+//! - Max Fierro, 12/3/2023 (maxfierro@berkeley.edu)
+
+use std::collections::HashMap;
+
+use crate::game::{Acyclic, Bounded, DTransition, STransition};
+use crate::interface::IOMode;
+use crate::model::{PlayerCount, State};
+use crate::solver::MAX_TRANSITIONS;
+
+/* SOLVER INTERFACES */
+
+/// Provides behavior for strongly solving games whose states do not repeat
+/// during play. A strong solution inherently visits all positions in the
+/// game, determines the best possible strategy for the player whose turn it
+/// is at that position, and associates the value achievable through that
+/// strategy to the position. This interface allows for an arbitrary number
+/// of state transitions being possible from all game states by using a
+/// dynamically-sized data structure at the cost of performance.
+pub trait DynamicSolver<const N: PlayerCount, S>
+where
+    Self: Acyclic<N> + DTransition<S> + Bounded<S>,
+{
+    /// Has the side effect of writing a strong solution to the underlying
+    /// game by exploring using `from` as a starting state. For specifics
+    /// on when files are written, see `cli::IOMode`. This function assumes
+    /// that `from` is a valid state encoding of the underlying game.
+    fn solve(&self, mode: IOMode);
+}
+
+/// Provides behavior for strongly solving games whose states do not repeat
+/// during play. A strong solution inherently visits all positions in the
+/// game, determines the best possible strategy for the player whose turn it
+/// is at that position, and associates the value achievable through that
+/// strategy to the position. This interface specifies that the traversal
+/// of the game must be possible via stack-allocated data structures by
+/// requiring the implementation of `STransition` with a preset number
+/// of maximum state transitions `solvers::MAX_TRANSITIONS`.
+pub trait StaticSolver<const N: PlayerCount, const F: usize, S>
+where
+    Self: Acyclic<N> + STransition<S, F> + Bounded<S>,
+{
+    /// Has the side effect of writing a strong solution to the underlying
+    /// game by exploring using `from` as a starting state. For specifics
+    /// on when files are written, see `cli::IOMode`. This function assumes
+    /// that `from` is a valid state encoding of the underlying game.
+    fn solve(&self, mode: IOMode);
+}
+
+/* BLANKET IMPLEMENTATIONS */
+
+impl<const N: PlayerCount, G> DynamicSolver<N, State> for G
+where
+    G: Acyclic<N> + DTransition<State> + Bounded<State>,
+{
+    fn solve(&self, mode: IOMode) {
+        let mut db: HashMap<State, Record<N>> = HashMap::new();
+        dynamic_backward_induction(&mut db, self);
+    }
+}
+
+impl<const N: PlayerCount, G> StaticSolver<N, MAX_TRANSITIONS, State> for G
+where
+    G: Acyclic<N> + STransition<State, MAX_TRANSITIONS> + Bounded<State>,
+{
+    fn solve(&self, mode: IOMode) {
+        let mut db = BPDatabase::new(self.id(), mode);
+        static_backward_induction(&mut db, self);
+    }
+}
+
+/* SOLVING ALGORITHMS */
+
+fn dynamic_backward_induction<const N: PlayerCount, G>(
+    db: &mut BPDatabase<N>,
+    game: &G,
+) where
+    G: Bounded<State> + DTransition<State>,
+{
+    let mut stack = Vec::new();
+    stack.push(game.start());
+    while let Some(curr) = stack.pop() {
+        let children = game.prograde(curr);
+        if let None = db.get(curr) {
+            db.put(curr, Record::default());
+            if children.is_empty() {
+                let record = Record::default()
+                    .with_utility(game.utility(curr))
+                    .with_remoteness(0);
+                db.put(curr, record)
+            } else {
+                stack.push(curr);
+                stack.extend(
+                    children
+                        .iter()
+                        .filter(|&x| db.get(*x).is_none()),
+                );
+            }
+        } else {
+            db.put(
+                curr,
+                children
+                    .iter()
+                    .map(|&x| db.get(x).unwrap())
+                    .max_by(|r1, r2| r1.cmp(&r2, game.turn(curr)))
+                    .unwrap(),
+            );
+        }
+    }
+}
+
+fn static_backward_induction<const N: PlayerCount, G>(
+    db: &mut BPDatabase<N>,
+    game: &G,
+) where
+    G: StaticSolver<N, MAX_TRANSITIONS, State> + Bounded<State>,
+{
+    let mut stack = Vec::new();
+    stack.push(game.start());
+    while let Some(curr) = stack.pop() {
+        let children = game.transition(curr);
+        if let None = db.get(curr) {
+            db.put(curr, Record::default());
+            if children
+                .iter()
+                .all(|x| x.is_none())
+            {
+                let record = Record::default()
+                    .with_utility(game.utility(curr))
+                    .with_remoteness(0);
+                db.put(curr, record)
+            } else {
+                stack.push(curr);
+                stack.extend(
+                    children
+                        .iter()
+                        .filter_map(|&x| x)
+                        .filter(|&x| db.get(x).is_none()),
+                );
+            }
+        } else {
+            db.put(
+                curr,
+                children
+                    .iter()
+                    .filter_map(|&x| x)
+                    .map(|x| db.get(x).unwrap())
+                    .max_by(|r1, r2| r1.cmp(&r2, game.turn(curr)))
+                    .unwrap(),
+            );
+        }
+    }
+}
