+++ src/solver/strong/cyclic.rs
@@ -1,0 +1,9 @@
+//! # Strong Cyclic Solving Module
+//!
+//! This module implements strong cyclic solvers for all applicable types of
+//! games through blanket implementations of the `acyclic::Solver` trait,
+//! optimizing for specific game characteristics wherever possible.
+//!
+//! #### Authorship
+//!
+//! - Max Fierro, 12/3/2023 (maxfierro@berkeley.edu)
