+++ src/solver/weak/acyclic.rs
@@ -1,0 +1,9 @@
+//! # Weak Acyclic Solving Module
+//!
+//! This module implements weak acyclic solvers for all applicable types of
+//! games through blanket implementations of the `acyclic::Solver` trait,
+//! optimizing for specific game characteristics wherever possible.
+//!
+//! #### Authorship
+//!
+//! - Max Fierro, 12/4/2023 (maxfierro@berkeley.edu)
